refactor: split mail logic into modules
