+++ src/mail/mod.rs
@@ -1,0 +1,3 @@
+pub mod auth;
+pub mod storage;
+pub mod imap;
